tidy up header #1
